Makefile compile en cpp

+++ Day00/ex01/PhoneBook.cpp
@@ -13,7 +13,7 @@ int			PhoneBook::setfirst_name(std::string str)
 	{
 		if (std::isalpha(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_first_name = str;
 			return (1);
@@ -29,7 +29,7 @@ int			PhoneBook::setlast_name(std::string str)
 	{
 		if (std::isalpha(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_last_name = str;
 			return (1);
@@ -44,7 +44,7 @@ int			PhoneBook::setnickname(std::string str)
 	{
 		if (std::isalpha(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_nickname = str;
 			return (1);
@@ -59,7 +59,7 @@ int			PhoneBook::setlogin(std::string str)
 	{
 		if (std::isalnum(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_login = str;
 			return (1);
@@ -75,7 +75,7 @@ int			PhoneBook::setpostal_address(std::string str)
 		if (std::isalnum(str[i]) != 1 && str[i] != ' ' && str[i] != '-' &&
 			str[i] != '\'' && str[i] != ',')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_postal_address = str;
 			return (1);
@@ -90,7 +90,7 @@ int			PhoneBook::setemail_address(std::string str)
 	{
 		if (std::isprint(str[i]) != 1)
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_email_address = str;
 			return (1);
@@ -105,7 +105,7 @@ int			PhoneBook::setphone_number(std::string str)
 	{
 		if (std::isdigit(str[i]) != 1 && str[i] != ' ')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_phone_number = str;
 			return (1);
@@ -120,7 +120,7 @@ int			PhoneBook::setbirthday_date(std::string str)
 	{
 		if (std::isdigit(str[i]) != 1 && str[i] != ' ' && str[i] != '/' && str[i] != '.')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_birthday_date = str;
 			return (1);
@@ -135,7 +135,7 @@ int			PhoneBook::setfavorite_meal(std::string str)
 	{
 		if (std::isalpha(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_favorite_meal = str;
 			return (1);
@@ -150,7 +150,7 @@ int			PhoneBook::setunderwear_color(std::string str)
 	{
 		if (std::isalpha(str[i]) != 1 && str[i] != ' ' && str[i] != '-')
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_underwear_color = str;
 			return (1);
@@ -165,7 +165,7 @@ int			PhoneBook::setdarkest_secret(std::string str)
 	{
 		if (std::isprint(str[i]) != 1)
 			return (0);
-		if (i + 1 == str.size())
+		if (i + 1 == (int)str.size())
 		{
 			this->_darkest_secret = str;
 			return (1);
